// helpers/types.ts
export enum eFlowNetwork {
  main = "flow",
  testnet = "flow-testnet",
}
// ...
export interface IBaseConfiguration {
  MarketId: string;
  ATokenNamePrefix: string;
  StableDebtTokenNamePrefix: string;
  VariableDebtTokenNamePrefix: string;
  SymbolPrefix: string;
  ProviderId: number;
  TestnetMarket?: boolean;
  ProviderRegistryOwner?: iParamsPerNetwork<tEthereumAddress | undefined>;
  FallbackOracle?: iParamsPerNetwork<tEthereumAddress>;
  ChainlinkAggregator: iParamsPerNetwork<ITokenAddress>;
  WrappedTokenGateway?: iParamsPerNetwork<tEthereumAddress>;
  ReserveFactorTreasuryAddress: iParamsPerNetwork<tEthereumAddress>;
  StableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  VariableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  ReserveAssets?: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  OracleQuoteCurrency: string;
  OracleQuoteUnit: string;
  OracleQuoteCurrencyAddress: tEthereumAddress;
  ReservesConfig: SymbolMap<IReserveParams>;
  WrappedNativeTokenSymbol: string; // ex: WETH or WMATIC
  IncentivesConfig: IncentivesConfig;
  EModes: SymbolMap<EMode>;
  L2PoolEnabled?: iParamsPerNetwork<boolean>;
  StkAaveProxy?: iParamsPerNetwork<tEthereumAddress>;
  ParaswapRegistry?: iParamsPerNetwork<tEthereumAddress>;
  FlashLoanPremiums: {
    total: number;
    protocol: number;
  };
  RateStrategies: IStrategy;
}

export interface ICommonConfiguration extends IBaseConfiguration {}

export interface IAaveConfiguration extends ICommonConfiguration {
  //ReservesConfig: iAavePoolAssets<IReserveParams>;
}